feat(DTFS2-7393): revert facility/fed-disabled directory
