+++ e2e-tests/gef/cypress/e2e/facility/version-0/version-0-facility.spec.js
@@ -1,8 +1,8 @@
-import relative from '../../../relativeURL';
-import { BANK1_MAKER1 } from '../../../../../../e2e-fixtures/portal-users.fixture';
-import applicationDetails from '../../../pages/application-details';
-import facilities from '../../../pages/facilities';
-import aboutFacility from '../../../pages/about-facility';
+import relative from '../../relativeURL';
+import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
+import applicationDetails from '../../pages/application-details';
+import facilities from '../../pages/facilities';
+import aboutFacility from '../../pages/about-facility';
 
 /**
  * NOTE: These tests check the backwards compatibility with in-flight version 0 deals.
