+++ e2e-tests/gef/cypress/e2e/facility/version-0/version-0-unissued-facility.spec.js
@@ -1,8 +1,8 @@
-import relative from '../../../relativeURL';
-import { BANK1_MAKER1 } from '../../../../../../e2e-fixtures/portal-users.fixture';
+import relative from '../../relativeURL';
+import { BANK1_MAKER1 } from '../../../../../e2e-fixtures/portal-users.fixture';
 
-import aboutFacilityUnissued from '../../../pages/unissued-facilities-about-facility';
-import { DEAL_STATUS } from '../../../../fixtures/constants';
+import aboutFacilityUnissued from '../../pages/unissued-facilities-about-facility';
+import { DEAL_STATUS } from '../../../fixtures/constants';
 
 /**
  * NOTE: These tests check the backwards compatibility with in-flight version 0 deals.
